+++ client/app/modules/home/home.routes.js
@@ -2,7 +2,7 @@ export default function routes($routeProvider) {
   'ngInject';
   $routeProvider.when('/', {
     template: '<home></home>',
-    title: 'Schoolhouse Yoga',
+    title: 'Schoolhouse Yoga Pittsburgh: Squirrel Hill, East Liberty, Ross Park',
     resolve: {
       '': (AnnouncementService, HomeService) => Promise.all([AnnouncementService.initialized, HomeService.initialized])
     }