+++ client/app/modules/locations/locations.routes.js
@@ -2,7 +2,7 @@ export default function($routeProvider) {
   'ngInject';
   $routeProvider.when('/locations', {
     template: '<locations></locations>',
-    title: 'Schoolhouse Yoga Locations',
+    title: 'Schoolhouse Yoga Pittsburgh Locations',
     resolve: {
       '': LocationService => LocationService.initialized
     }