+++ client/app/modules/teachers/teachers.routes.js
@@ -2,7 +2,7 @@ export default function($routeProvider) {
   'ngInject';
   $routeProvider.when('/teachers', {
     template: '<teachers></teachers>',
-    title: 'Schoolhouse Yoga Teachers',
+    title: 'Schoolhouse Yoga Pittburgh\'s Teachers',
     resolve: {
       '': TeacherService => TeacherService.initialized
     }