+++ client/app/modules/teachertraining/teachertraining.routes.js
@@ -2,7 +2,7 @@ export default function($routeProvider) {
   'ngInject';
   $routeProvider.when('/teachertraining', {
     template: '<teachertraining></teachertraining>',
-    title: 'Pittsburgh Yoga Teacher Training - 200 Hour Program',
+    title: 'Pittsburgh Yoga Teacher Training - 200 Hour Program - Yoga Alliance Certified',
     resolve: {
       '': Cart => Cart.initialized
     }