+++ client/app/modules/workshops/workshops.routes.js
@@ -2,7 +2,7 @@ export default function($routeProvider) {
   'ngInject';
   $routeProvider.when('/workshops', {
     template: '<workshops></workshops>',
-    title: 'Schoolhouse Yoga Workshops',
+    title: 'Schoolhouse Yoga Pittsburgh\'s Workshops',
     resolve: {
       '': (WorkshopService, Cart) => Promise.all([WorkshopService.initialized, Cart.initialized])
     }